+++ client/src/components/PageFormatter.jsx
@@ -1,7 +1,7 @@
 import { Outlet } from "react-router-dom";
 import { useAppSelector } from "../redux/store";
 import { NavLink } from "react-router-dom";
-import AuthNavBar from "./AuthNavBar";
+import AuthNavBar from "./Auth/AuthNavBar";
 const PageFormatter = () => {
   const { user } = useAppSelector((state) => state.auth);
   if (user == null) {
