+++ client/src/pages/LeaderboardsPage.jsx
@@ -9,7 +9,7 @@ const LeaderboardsPage = () => {
     <div className="flex justify-center items-start h-screen w-screen bg-gradient-to-b from-color7 to-color3">
       {data ? (
         <div className='w-5/6 flex justify-center flex-col'>
-          <h1 className="text-3xl font-bold text-center mb-3">Leaderboards</h1>
+          <h1 className="text-3xl font-bold text-center mb-3 dark:text-red-400">Leaderboards</h1>
           <LeaderBoard users={data} />
         </div>
       ) : isLoading ? (
